+++ torchvision/prototype/models/__init__.py
@@ -1,3 +1,2 @@
-from . import optical_flow
 from . import segmentation
 from . import video
