--- torchvision/prototype/models/optical_flow/__init__.py
@@ -1,1 +1,0 @@
-from .raft import RAFT, raft_large, raft_small, Raft_Large_Weights, Raft_Small_Weights
